test: pkg/mod module
